+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Collection: optional support for Samsung and Sony burst patterns
 - Info: improved state/place display (requires rescan, limited to AU/GB/EN)
 - improved support for system font scale